Multiple Components in Sync,Lifting the State Up,Stateless Functional Components
