+++ src/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.css'
-import Counters from './component/listCounter'
-
+import App from "./App"
 ReactDOM.render(
-    <Counters/>,
+    <App/>,
     document.getElementById('root'));
